Memory game
Memory Game by Mehmood Shahid Awan
A fun and interactive Memory game built with HTML, CSS, and JavaScript. The game challenges players to match pairs of cards with identical images. It features a timer, a scoring system, and a reset button. Perfect for improving memory and practicing JavaScript event handling and DOM manipulation.